requirements.txt
torch==2.6.0
transformers==4.46.3
tokenizers==0.20.3
einops
addict 
easydict
setuptools
psutil
pillow 
pdf2image
opencv-python
torchvision
matplotlib
